HTML + CSS and a bit of JS
Made a more interactive page for a more streamlined one page look.

When you click on the nav bar about US and How to Play it makes it so a side container slides out from the outside of the page, and gives out the information.

removed the necessary HTML + CSS to make this look work, however, there's some issues implementing it with the current HTML and it causes crashes